feat: add super-linter support (#5)
This adds support for the super-linter to the project. This provides
linting for a myriad of languages and formats. This also adds a
pr-checks workflow to run the linter on pull requests. A script has been
added to run the linter locally. The script will run the linter on all
files in the project.

Refs: #3
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>