feat: cache (dev)container builds (#22)
This refactors the devcontainer build to separate the texlive base
container to its own directory. This allows us to cache the texlive
container build and only rebuild it as needed.

This also adds new workflows to build the texlive container and the
devcontainer during PRs when the relevant files change. A new composite
action was added to abstract the container build process.

The release-please workflow was updated to release the container and
devcontainer images to GHCR. The versions of the images are grouped to
ensure that one is revbumbed when the other is.

Refs: #19
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>