fix: ignore CHANGELOGs in PRs and container builds (#36)
This adds an exclusion to the pr-devcontainer and pr-latex-container
workflows to ignore changes to the CHANGELOGs. This is to prevent
unnecessary rebuilds of the containers, slowing the PR process.

.dockerignore files were also added to the devcontainer and latex base
containers to prevent the CHANGELOGs from being copied into the build
context. The devcontainer .dockerignore file was also updated to ignore
the devcontainer-specific files.

Refs: #35
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>