feat: implement docker caching and semver tagging
This updates the container action to use docker caching. This required
updating the action to require registry access for all usage.

The tagging logic was updated to enforce semver tags and take the
version components in as inputs.

The resulting image and tags are loaded into the environment for use
even if they are not being pushed.

Refs: #48
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>