feat: configure eslint to parse JSONC files
This adds an .eslintrc.json file to the root of the project. This file
is configured to use the eslint-plugin-jsonc plugin to parse .json
files as JSONC.

This also adds a .gitignore file to the root of the project to ignore
the node_modules directory. This was needed becasue the
eslint-plugin-jsonc is a dependency of the project now and must be
installed for super-linter to work.

The pr-checks workflow was updated to ensure that dependencies are
installed before executing the linting. The lint script was also updated
to ensure the same thing.

Refs: #54
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>