feature: devcontainer publishing
This moves the devcontainer to a subdirectory. The release-please config
and workflow was updated to use this new location. The published image
name was changed to latex-devcontainer. The CHANGELOG was also moved.

The pr-devcontainer workflow was updated to use the new location.

A new decontainer config was created in the original location which
references the published image.

Refs: #46
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>