fix: load and push devcontainer as separate steps
This fixes the issue where the decontainer builds would fail to tag
after building. For an unknown reason, the build would fail to load the
image into the docker image store. This caused subsequent tagging to
fail.

This fixes the issue by building the devcontainer and outputting it to a
tar file. If the devcontainer is to be pushed, the tar is loaded and
then tagged before being pushed.

Refs: #31
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>