feat: add DevOps Engineer Resume, CV, and utils
This introduces the utilities and templates for generating a DevOps
Engineer resume and CV by rendering LaTeX templates with Jinja2, using
data from the nearby YAML files. This approach allows for easy
customization and management of the content, making it simple to
tailor the documents to specific job applications.

Linting rules were updated to ensure compatibility with the new files.

Fixes: #144
Signed-off-by: Jaremy Hatler <root@jhatler.com>